Añadido query GetHabitacionByID

+++ src/main/java/com/hotel/Continental/controller/HabitacionController.java
@@ -40,7 +40,21 @@ public class HabitacionController {
 
     @DeleteMapping(value = "/delete")
     public int deleteHabitacion(@RequestBody HabitacionDto habitacionDto){
+        if (habitacionDto.getIdHabitacion() == null) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "IdHabitacion is required");
+        }
+        if(getHabitacionById(habitacionDto.getIdHabitacion()) == null){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Habitacion not found");
+        }
         return habitacionService.deleteHabitacion(habitacionDto);
+    }
+    @GetMapping(value = "/getHabitacionById/{idHabitacion}")
+    public HabitacionDto getHabitacionById(@PathVariable int idHabitacion) {
+        HabitacionDto habitacionDto = habitacionService.getHabitacionById(idHabitacion);
+        if (habitacionDto == null){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Habitacion not found");
+        }
+        return habitacionService.getHabitacionById(idHabitacion);
     }
 }
